Create epidemic (optimized).c
after giving this my best attempt, it only sped up the best case scenario by ~2x. However, when N is small, the improvements are several hundredfold. However, I'm pretty sure, that if one used just an array of structs with SIR pairs, and just incremented/decremented those counts, and (critically) ignored the same random update order that is used here, there would be greater improvements to be had. But this is the most I was willing to put into this homework assignment.